commit for test cases

+++ currency/src/main/java/com/kpi/microservice/currencyexchangeexampleservice/demo2/CurrencyConversionSampleBean.java
@@ -1,21 +1,20 @@
 package com.kpi.microservice.currencyexchangeexampleservice.demo2;
 
-import java.math.BigDecimal;
 
 public class CurrencyConversionSampleBean {
 	private long id;
 	private String from;
 	private String to;
-	private BigDecimal quantity;
-	private BigDecimal ConversionMultiple;
-	private BigDecimal totalCalculatedAmount;
+	private int quantity;
+	private int ConversionMultiple;
+	private int totalCalculatedAmount;
 	private int port;
 	
 	public CurrencyConversionSampleBean () {
 		
 	}
 	
-	public CurrencyConversionSampleBean (Long id, String from, String to, BigDecimal conversionMultiple, BigDecimal quantity, BigDecimal totalCalculatedAmount, int port) {
+	public CurrencyConversionSampleBean (Long id, String from, String to, int conversionMultiple, int quantity, int totalCalculatedAmount, int port) {
 		super();  
 		this.id = id;  
 		this.from = from;  
@@ -44,22 +43,22 @@ public class CurrencyConversionSampleBean {
 	public void setTo(String to) {
 		this.to = to;
 	}
-	public BigDecimal getQuantity() {
+	public int getQuantity() {
 		return quantity;
 	}
-	public void setQuantity(BigDecimal quantity) {
+	public void setQuantity(int quantity) {
 		this.quantity = quantity;
 	}
-	public BigDecimal getConversionMultiple() {
+	public int getConversionMultiple() {
 		return ConversionMultiple;
 	}
-	public void setConversionMultiple(BigDecimal conversionMultiple) {
+	public void setConversionMultiple(int conversionMultiple) {
 		ConversionMultiple = conversionMultiple;
 	}
-	public BigDecimal getTotalCalculatedAmount() {
+	public int getTotalCalculatedAmount() {
 		return totalCalculatedAmount;
 	}
-	public void setTotalCalculatedAmount(BigDecimal totalCalculatedAmount) {
+	public void setTotalCalculatedAmount(int totalCalculatedAmount) {
 		this.totalCalculatedAmount = totalCalculatedAmount;
 	}
 	public int getPort() {
